+++ lib/src/user32/get_cursor_pos.dart
@@ -1,23 +1,6 @@
 import 'dart:ffi';
 
-import 'package:ffi/ffi.dart';
 import 'user32.dart';
-
-/// Point struct
-class Point extends Struct {
-  /// Horizontal coordinate.
-  @Int32()
-  int x;
-
-  /// Vertical coordinate.
-  @Int32()
-  int y;
-
-  /// Allocate a new point.
-  factory Point.allocate({int x = 0, int y = 0}) => allocate<Point>().ref
-    ..x = x
-    ..y = y;
-}
 
 typedef GetCursorC = Uint8 Function(Pointer<Point> point);
 
@@ -25,7 +8,7 @@ typedef GetCursorDart = int Function(Pointer<Point> point);
 
 /// Get the current cursor position.
 /// See https://docs.microsoft.com/en-us/windows/win32/api/winuser/nf-winuser-getcursorpos
-bool GetCursorPos(Point point) {
+bool GetCursorPos({@required Point point}) {
   final GetCursorP =
       dylib.lookupFunction<GetCursorC, GetCursorDart>('GetCursorPos');