+++ lib/src/user32/get_window_text_length.dart
@@ -16,8 +16,8 @@ typedef GetWindowTextLengthDart = int Function(Pointer<Hwnd> hWnd);
 /// Retrieves the length, in characters, of the specified window's title bar
 /// text (if the window has a title bar).
 /// See https://docs.microsoft.com/en-us/windows/win32/api/winuser/nf-winuser-getwindowtextlengtha
-int GetWindowTextLength(Pointer<Hwnd> hWnd,
-    {TextFormat textFormat = TextFormat.utf16}) {
+int GetWindowTextLength(
+    {@required Pointer<Hwnd> hWnd, TextFormat textFormat = TextFormat.utf16}) {
   var symbol = textFormat == TextFormat.utf16
       ? 'GetWindowTextLengthW'
       : 'GetWindowTextLengthA';