+++ lib/src/user32/hwnd.dart
@@ -1,4 +1,0 @@
-import 'dart:ffi';
-
-/// Handle type.
-class Hwnd extends Struct {}