+++ lib/src/user32/set_cursor_position.dart
@@ -14,10 +14,15 @@ typedef SetCursorPosDart = int Function(
 
 /// Set the cursor position
 /// See https://docs.microsoft.com/en-us/windows/win32/api/winuser/nf-winuser-setcursorpos
-bool SetCursorPos(int x, int y) {
+bool SetCursorPos({@required int x, @required int y}) {
   final SetCursorPosP =
       dylib.lookupFunction<SetCursorPosC, SetCursorPosDart>('SetCursorPos');
 
   var result = SetCursorPosP(x, y);
   return result != 0;
 }
+
+/// Set the cursor position given a point structure.
+/// See https://docs.microsoft.com/en-us/windows/win32/api/winuser/nf-winuser-setcursorpos
+bool SetCursorPosFromPoint({Point point}) =>
+    SetCursorPos(x: point.x, y: point.y);