+++ lib/src/user32/set_windows_hook_ex.dart
@@ -4,6 +4,7 @@ import 'package:meta/meta.dart';
 
 import 'user32.dart';
 
+//
 /*
 HHOOK SetWindowsHookExA(
   int       idHook,
@@ -13,14 +14,14 @@ HHOOK SetWindowsHookExA(
 );
  */
 
-typedef SetWindowsHookExAC = Pointer Function(
+typedef SetWindowsHookExAC = Pointer<HHook> Function(
   Uint32 code,
   Pointer<NativeFunction<HookprocC>> lpfn,
   Pointer hmod,
   Uint32 dwThreadId,
 );
 
-typedef SetWindowsHookExADart = Pointer Function(
+typedef SetWindowsHookExADart = Pointer<HHook> Function(
   int code,
   Pointer<NativeFunction<HookprocC>> lpfn,
   Pointer hmod,
@@ -51,19 +52,23 @@ const _expReturn = -1;
 /// Installs an application-defined hook procedure into a hook chain.
 /// See https://docs.microsoft.com/en-us/windows/win32/api/winuser/nf-winuser-setwindowshookexa
 @experimental
-Pointer SetWindowsHookExA(int idHook, HookprocDart callback,
-    {Pointer hmod,
+@alwaysThrows
+Pointer<HHook> SetWindowsHookEx(
+    {@required int idHook,
+    @required HookprocDart callback,
+    Pointer hmod,
     int dwThreadId = 0,
     TextFormat textFormat = TextFormat.utf16}) {
-  var symbol = textFormat == TextFormat.utf16
+  throw UnimplementedError('SetWindowsHookEx is not yet implemented');
+/*  var symbol = textFormat == TextFormat.utf16
       ? 'SetWindowsHookExW'
       : 'SetWindowsHookExA';
   hmod ??= nullptr;
   var cb;
 //  var cb = Pointer.fromFunction<HookprocC>(callback, _expReturn);
-  final SetWindowsHookExAP =
+  final SetWindowsHookExP =
       dylib.lookupFunction<SetWindowsHookExAC, SetWindowsHookExADart>(symbol);
 
-  var result = SetWindowsHookExAP(idHook, cb, hmod, dwThreadId);
-  return result;
+  var result = SetWindowsHookExP(idHook, cb, hmod, dwThreadId);
+  return result;*/
 }