+++ lib/src/user32/user32.dart
@@ -1,4 +1,5 @@
 import 'dart:ffi';
+export 'package:meta/meta.dart';
 
 export 'call_next_hook_ex.dart';
 export 'find_window_ex.dart';
@@ -8,18 +9,13 @@ export 'get_key_state.dart';
 export 'get_message.dart';
 export 'get_window_text.dart';
 export 'get_window_text_length.dart';
-export 'hwnd.dart';
 export 'keybd_event.dart';
 export 'mouse_event.dart';
 export 'register_hotkey.dart';
 export 'send_input.dart';
 export 'set_cursor_position.dart';
 export 'set_windows_hook_ex.dart';
+export 'types.dart';
 
 /// User32.dll library
 final dylib = DynamicLibrary.open('User32.dll');
-
-enum TextFormat {
-  ansi,
-  utf16
-}